requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dateutil==2.8.2
pytz==2023.3
pytest==7.4.0
pytest-cov==4.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0